pyproject.toml
[project]
name = "tree-enhanced-codebert"
version = "1.0.0"
description = "Default template for PDM package"
authors = [
    {name = "Patryk Bartkowiak", email = "patrykbart99@gmail.com"},
]
dependencies = [
    "gdown==5.2.0",
    "ipykernel==6.29.5",
    "datasets==3.2.0",
    "ipywidgets==8.1.5",
    "transformers[torch]==4.48.1",
    "docopt==0.6.2",
    "dpu-utils==0.6.1",
    "tree-sitter==0.23.1",
    "tree-sitter-python==0.23.4",
    "tree-sitter-java==0.23.5",
    "tree-sitter-php==0.23.11",
    "tree-sitter-ruby==0.23.1",
    "tree-sitter-javascript==0.23.1",
    "tree-sitter-go==0.23.4",
    "wandb==0.19.4",
    "torch==2.5.1",
    "huggingface-hub==0.26.0",
    "pyyaml==6.0.2",
    "scikit-learn==1.6.1",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}

[[tool.pdm.source]]
url = "https://download.pytorch.org/whl/cu124"
verify_ssl = true
include_packages = ["torch", "torchvision", "torchaudio"]
name = "cuda"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = false

[tool.pdm.scripts]
parse_data = {cmd = "src/parse_dataset.py"}
train = {cmd = "src/train.py"}
eval_pretrained = {cmd = "src/eval_pretrained.py"}
finetune = {cmd = "src/finetune_clone_detection.py"}
